+++ HomePage/index.html
@@ -38,9 +38,6 @@
 <!-- Sign Up Section --> 
 <!-- Instrument Selector -->
 <!-- Genre Selector -->
-
-
-
     <script>
       let croud = document.getElementById('croud');
       let note = document.getElementById('note');
