+++ controllers/api/userRoutes.js
@@ -1,5 +1,11 @@
 const router = require('express').Router();
 const { User } = require('../../models');
+const Connection = require('./models/Connections.js');
+const Genre = require('../models/Genre');
+const Instrument = require('../models/Instrument');
+
+
+
 
 router.post('/', async (req, res) => {
   try {
@@ -47,6 +53,97 @@ router.post('/login', async (req, res) => {
     res.status(400).json(err);
   }
 });
+
+router.get('/user/:id', async (req, res) => {
+  try {
+    const projectData = await Project.findByPk(req.params.id, {
+      include: [
+        {
+          model: User,
+          attributes: ['name','user_instrument','user_genre','user_id'],
+        },
+      ],
+    });
+
+    const project = projectData.get({ plain: true });
+
+    res.render('project', {
+      ...project,
+      logged_in: req.session.logged_in
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+router.get('./Instrument', async (req, res) => {
+  try {
+    const projectData = await Project.findByPk(req.params.id, {
+      include: [
+        {
+          model: Instrument,
+          attributes: [],
+        },
+      ],
+    });
+
+    const project = projectData.get({ plain: true });
+
+    res.render('project', {
+      ...project,
+      logged_in: req.session.logged_in
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+router.get('./Genre', async (req, res) => {
+  try {
+    const projectData = await Project.findByPk(req.params.id, {
+      include: [
+        {
+          model: Genre,
+          attributes: [],
+        },
+      ],
+    });
+
+    const project = projectData.get({ plain: true });
+
+    res.render('project', {
+      ...project,
+      logged_in: req.session.logged_in
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+router.get('./Connections', async (req, res) => {
+  try {
+    const projectData = await Project.findByPk(req.params.id, {
+      include: [
+        {
+          model: Connection,
+          attributes: [],
+        },
+      ],
+    });
+
+    const project = projectData.get({ plain: true });
+
+    res.render('project', {
+      ...project,
+      logged_in: req.session.logged_in
+    });
+  } catch (err) {
+    res.status(500).json(err);
+  }
+});
+
+
+
 
 router.post('/logout', (req, res) => {
   if (req.session.logged_in) {
